+++ src/components/file-editor/FileEditor.constants.ts
@@ -1,0 +1,1 @@
+export const initialFileValue = '<p>Create your file here.</p>'
