+++ src/constants/translations/en/breadcrumbs.json
@@ -13,6 +13,5 @@
   "newLesson": "New lesson",
   "chat": "Chat",
   "myOffers": "My offers",
-  "newLesson":"New lesson",
   "myLesson":"My lesson"
 }
